Refactor codebase to rename http_cli to PyFetch, update package structure.
