+++ PyFetch/__init__.py
@@ -1,6 +1,6 @@
 """HTTP CLI client library for making HTTP requests."""
 
-from http_cli.exceptions import HTTPClientError, HTTPConnectionError, ResponseError
-from http_cli.http_client import HTTPClient
+from PyFetch.exceptions import HTTPClientError, HTTPConnectionError, ResponseError
+from PyFetch.http_client import HTTPClient
 
 __all__ = ["HTTPClient", "HTTPClientError", "HTTPConnectionError", "ResponseError"]
