+++ PyFetch/__main__.py
@@ -2,7 +2,7 @@
 
 import sys
 
-from http_cli.cli import main
+from PyFetch.cli import main
 
 if __name__ == "__main__":
     try:
