+++ PyFetch/exceptions.py
